Atualizando CRUD Usiários

+++ config_bd/bd_deletar_usuario.php
@@ -1,10 +1,19 @@
 <?php
 include ("../config_bd/conexao.php");
-
+session_start();
 $id = $_POST['id'];
+$acao = $_POST['acao'];
+$_SESSION['user_id_edit'] = $id;
 
-$sql1 = pg_query("DELETE FROM usuarios WHERE id='$id';") or die("Erro no comando SQL");
+if($acao == "Excluir"){
+   $sql1 = pg_query("DELETE FROM usuarios WHERE id='$id';") or die("Erro no comando SQL");
+   header("Location:../paginas/usuarios.php");
 
-header("Location:../paginas/usuarios.php");
+} else {
+  header("Location:../config_bd/../paginas/editar_usuario.php");
+  
+}
+
+
 
 ?>
